pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "phrasely"
version = "0.1.0"
description = "Phrase clustering and embedding pipeline (GPU default)"
authors = [
{ name = "Michael Mastroianni", email = "mmastroianni@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
"numpy",
"pandas",
"scikit-learn",
"hdbscan",
"sentence-transformers",
"tqdm",
"psutil",
"pyyaml",
"joblib"
]


[project.optional-dependencies]
dev = ["pytest", "flake8", "black", "mypy", "jupyterlab", "types-requests", "isort"]
gpu = ["cuml>=24.06", "cudf>=24.06", "cupy-cuda12x"]


[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length=88

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src", "tests"]
known_first_party = ["phrasely"]
combine_as_imports = true

[flake8]
max-line-length=88
